Module Cobol_preproc

module Src_overlay : sig ... end

A source overlay associates complex source code elements with virtual limits of lexing tokens that can be fed to parsers so as to tag ASTs with complex source locations.

module Src_lexing : sig ... end

module Text : sig ... end

module Text_printer : sig ... end

module Copybook : sig ... end

type preprocessor

type input =

| Filename of string

| String of {

contents : string;

filename : string;

}

| Channel of {

contents : Stdlib.in_channel;

filename : string;

}

type init = {

init_libpath : string list;

init_config : Cobol_config.t;

init_source_format : Cobol_config.source_format_spec;

}

type log = log_entry list

and rev_log = log

and log_entry = {

matched_loc : Cobol_common.Srcloc.srcloc;

replacement_text : Text.text;

}

val diags : preprocessor -> Cobol_common.Diagnostics.Set.t

val add_diag : preprocessor -> Cobol_common.Diagnostics.t -> preprocessor

val add_diags : preprocessor -> Cobol_common.Diagnostics.Set.t -> preprocessor

val log : preprocessor -> log

val rev_log : preprocessor -> rev_log

val srclexer : preprocessor -> Cobol_preproc__.Preproc.any_srclexer

val position : preprocessor -> Stdlib.Lexing.position

val next_sentence : preprocessor -> Text.text * preprocessor

High-level commands

val decide_source_format : string -> Cobol_config.source_format_spec -> Cobol_config.source_format Cobol_common.Diagnostics.with_diags

val preprocessor : ?on_period_only:bool -> ?verbose:bool -> input -> [< `WithLibpath of init ] -> preprocessor

val lex_file : ?on_period_only:bool -> source_format:Cobol_config.source_format_spec -> ?ppf:Stdlib.Format.formatter -> ?epf:Stdlib.Format.formatter -> string -> unit

val fold_text_lines : ?on_period_only:bool -> source_format:Cobol_config.source_format_spec -> ?epf:Stdlib.Format.formatter -> (Text.text -> 'a -> 'a) -> string -> 'a -> 'a

val lex_lib : ?on_period_only:bool -> source_format:Cobol_config.source_format_spec -> libpath:string list -> ?ppf:Stdlib.Format.formatter -> ?epf:Stdlib.Format.formatter -> ([< `Alphanum | `Word ] * string) -> unit

val preprocess_file : source_format:Cobol_config.source_format_spec -> ?verbose:bool -> ?config:Cobol_config.t -> libpath:string list -> ?ppf:Stdlib.Format.formatter -> ?epf:Stdlib.Format.formatter -> string -> unit

val text_of_file : source_format:Cobol_config.source_format_spec -> ?verbose:bool -> ?config:Cobol_config.t -> libpath:string list -> ?epf:Stdlib.Format.formatter -> string -> Text.text

val text_of_input : source_format:Cobol_config.source_format_spec -> ?verbose:bool -> ?config:Cobol_config.t -> libpath:string list -> ?epf:Stdlib.Format.formatter -> input -> Text.text