Module Cobol_preproc.Env

Environment for preprocessing and compilation. Holds definitions of preprocessor constants (preprocessor DEFINES, computation process variables), along with compilation variables (78-level constants).

Accepted values for these preprocessor variables are:

Alphanumerics;

Booleans;

Fixed-point numerics (no plain Integer, nor floating-points).

For now, the set of values for compilation variables is equivalent to that of preprocessor variables.

module VAL = Cobol_data.Value

module VAR : sig ... end

Utility module that maps any string to a physically unique upper-cased internal representation.

module MAP : sig ... end

module TYPES : sig ... end

include module type of struct include TYPES end

type env = TYPES.env = {

preproc_vars : preproc_var_definition MAP.t;

compil_vars : compilation_var_definition MAP.t;

}

and var = VAR.t

and preproc_var_definition = compilation_variable_definition Cobol_common.Srcloc.TYPES.with_src

and compilation_var_definition = compilation_variable_definition Cobol_common.Srcloc.TYPES.with_src

and compilation_variable_definition = TYPES.compilation_variable_definition = {

compvar : var;

compvar_value : value Cobol_common.Srcloc.TYPES.with_src;

}

and value = TYPES.value =

| Alphanum of Cobol_data.Types.alphanum_value

| Boolean of Cobol_data.Types.boolean_value

| Numeric of Cobol_data.Types.fixed_value

type var_definition = TYPES.var_definition =

| Preproc_var of preproc_var_definition

| Compilation_var of compilation_var_definition

type lookup_error = TYPES.lookup_error =

| Undefined

exception REDEFINITION of {

prev_def_src : Cobol_common.Srcloc.TYPES.src;

}

type t = env

val pp_value : Stdlib.Format.formatter -> value -> unit

val pp_compilation_variable_definition : Stdlib.Format.formatter -> compilation_variable_definition -> unit

val pp : t Pretty.printer

val empty : env

val var : string -> var

val var' : string Cobol_common.Srcloc.TYPES.with_loc -> var Cobol_common.Srcloc.TYPES.with_loc

val mem_preproc_var : MAP.key -> env -> bool

val mem_preproc_var' : MAP.key Cobol_common__Srcloc.TYPES.with_loc -> env -> bool

val mem_compil_var : MAP.key -> env -> bool

val mem_compil_var' : MAP.key Cobol_common__Srcloc.TYPES.with_loc -> env -> bool

val mem_var : MAP.key -> env -> bool

val mem_var' : MAP.key Cobol_common__Srcloc.TYPES.with_loc -> env -> bool

val var_definition_of : var:MAP.key Cobol_common__Srcloc.TYPES.with_loc -> ?try_compil_vars:bool -> env -> (var_definition, lookup_error) Stdlib.result

val register_preproc_var : src:Cobol_common.Srcloc.src -> var -> value Cobol_common.Srcloc.TYPES.with_src -> env -> env * compilation_variable_definition Cobol_common.Srcloc.with_src

val define_preproc_var : loc:Cobol_common.Srcloc.TYPES.srcloc -> MAP.key Cobol_common__Srcloc.TYPES.with_loc -> value Cobol_common.Srcloc.TYPES.with_src -> ?override:bool -> t -> t * preproc_var_definition

val define_process_parameter : var -> value Cobol_common.Srcloc.TYPES.with_src -> t -> env * compilation_variable_definition Cobol_common.Srcloc.with_src

val undefine_preproc_var : MAP.key Cobol_common__Srcloc.TYPES.with_loc -> t -> t

val define_compilation_var : loc:Cobol_common.Srcloc.srcloc -> var Cobol_common__Srcloc.TYPES.with_loc -> value Cobol_common.Srcloc.TYPES.with_src -> t -> t * compilation_var_definition

val find_compilation_var : MAP.key -> env -> compilation_var_definition option

val alphanum_literal_value' : Cobol_data.Types.alphanum_literal Cobol_common.Srcloc.TYPES.with_loc -> value

val boolean_literal_value' : Cobol_data.Types.boolean_literal Cobol_common.Srcloc.TYPES.with_loc -> value

val numeric_literal_value' : Cobol_data.Types.fixed_literal Cobol_common.Srcloc.TYPES.with_loc -> value

val numeric_value' : Cobol_data.Types.fixed_value Cobol_common.Srcloc.TYPES.with_loc -> value

val alphanum_literal_value : Cobol_data.Types.alphanum_literal Cobol_common.Srcloc.TYPES.with_loc -> value Cobol_common.Srcloc.TYPES.with_src

val boolean_literal_value : Cobol_data.Types.boolean_literal Cobol_common.Srcloc.TYPES.with_loc -> value Cobol_common.Srcloc.TYPES.with_src

val numeric_literal_value : Cobol_data.Types.fixed_literal Cobol_common.Srcloc.TYPES.with_loc -> value Cobol_common.Srcloc.TYPES.with_src

val numeric_value : Cobol_data.Types.fixed_value Cobol_common.Srcloc.TYPES.with_loc -> value Cobol_common.Srcloc.TYPES.with_src