Module Cobol_preproc.Src_overlay

A source overlay associates complex source code elements with virtual limits of lexing tokens that can be fed to parsers so as to tag ASTs with complex source locations.

type limit = Stdlib.Lexing.position

Source overlay limits, that MUST be considered abstract and whose contents must not be inspected nor relied upon outside of this module. Such limits are to be fed to menhir-generated parsers, that expect lexing positons from the Stdlib.Lexing module, so unfortunately we can neither make this type abstract nor private.

module type MANAGER = sig ... end

Manager of source overlay limits. Includes some mutable state.

module New_manager (Id : sig ... end) () : MANAGER

Nanager module instantiation

val debug_oc : Stdlib.out_channel option Stdlib.ref