Module Src_overlay.New_manager

Nanager module instantiation

Parameters

module Id : sig ... end

Signature

val id : string

Identifier of the manager; may be used for debugging.

val limits : Cobol_common.Srcloc.srcloc -> limit * limit

limits loc creates and returns the left- and right-limit to the given location.

Note: if loc overlaps with locations that have already been fed to this function, then restart must first be called.

val link_limits : limit -> limit -> unit

link_limits left right links the right limit of a token t to the left limit of the subsequent token t' (that is to be fed to the parser right after t).

val join_limits : (limit * limit) -> Cobol_common.Srcloc.srcloc

join_limits (start_limit, end_limit) returns a source location that spans from start_limit to end_limit. start_limit must have been built (via limits) before end_limit.

val dummy_limit : limit

dummy_limit is a limit that may be fed to a parser or join_limits, but not given to link_limits.

val restart : unit -> unit

restart () instructs the manager that (a subset of) the limits previously created are now outdated and should not be relied upon.