Module Cobol_preproc

Source format

module Src_format : sig ... end

Text

"Text" refers to the source after manipulations by preprocessor statements.

type text = Text.text

type comments = Text.comments

module Text : sig ... end

module Text_printer : sig ... end

Miscellaneous support modules

module Src_overlay : sig ... end

A source overlay associates complex source code elements with virtual limits of lexing tokens that can be fed to parsers so as to tag ASTs with complex source locations.

module Copybook : sig ... end

module Trace : sig ... end

include module type of struct include Trace.TYPES end

type log_entry =

| FileCopy of {

copyloc : Cobol_common.srcloc;

status : copy_event_status;

}

| Replace of {

replloc : Cobol_common.srcloc;

}

| Replacement of {

matched_loc : Cobol_common.srcloc;

replacement_text : Text.text;

}

and copy_event_status =

| CopyDone of string

| CyclicCopy of string

| MissingCopy of Copybook.lib_not_found_info

type log

Main entry points for the processor itself

module Options : sig ... end

type preprocessor

type input =

| Filename of string

| String of {

contents : string;

filename : string;

}

| Channel of {

contents : Stdlib.in_channel;

filename : string;

}

val preprocessor : ?options:Cobol_preproc__.Preproc_options.preproc_options -> input -> preprocessor

val reset_preprocessor : ?new_position:Stdlib.Lexing.position -> preprocessor -> input -> preprocessor

val reset_preprocessor_for_string : string -> ?new_position:Stdlib.Lexing.position -> preprocessor -> preprocessor

val diags : preprocessor -> Cobol_common.Diagnostics.Set.t

val add_diag : preprocessor -> Cobol_common.Diagnostics.t -> preprocessor

val add_diags : preprocessor -> Cobol_common.Diagnostics.Set.t -> preprocessor

val log : preprocessor -> Cobol_preproc__.Preproc_trace.TYPES.log

val comments : preprocessor -> Text.comments

val position : preprocessor -> Stdlib.Lexing.position

val source_format : preprocessor -> Src_format.any

val newline_cnums : preprocessor -> int list

val next_chunk : preprocessor -> Text.text * preprocessor

High-level commands

val decide_source_format : string -> Cobol_config.source_format_spec -> Src_format.any Cobol_common.Diagnostics.with_diags

val lex_file : source_format:Cobol_config.source_format_spec -> ?ppf:Stdlib.Format.formatter -> ?epf:Stdlib.Format.formatter -> input -> unit

val lex_lib : source_format:Cobol_config.source_format_spec -> libpath:string list -> ?ppf:Stdlib.Format.formatter -> ?epf:Stdlib.Format.formatter -> ([< `Alphanum | `Word ] * string) -> unit

val fold_source_lines : source_format:Cobol_config.source_format_spec -> ?epf:Stdlib.Format.formatter -> f:(int -> Text.text -> 'a -> 'a) -> input -> 'a -> 'a

val preprocess_file : ?options:Cobol_preproc__.Preproc_options.preproc_options -> ?ppf:Stdlib.Format.formatter -> ?epf:Stdlib.Format.formatter -> string -> unit

val text_of_file : ?options:Cobol_preproc__.Preproc_options.preproc_options -> ?epf:Stdlib.Format.formatter -> string -> Text.text

val text_of_input : ?options:Cobol_preproc__.Preproc_options.preproc_options -> ?epf:Stdlib.Format.formatter -> input -> Text.text